PR-1: RL_Atari class makes more sense to be call from run_api.py
